pyproject.toml
[tool.poetry]
name = "DomainAdaptationForFactualConsistencyDetection"
version = "0.1.0"
description = "This package contains mechanisms for domain adaptation for the task of detecting factual consistency in text generated by large language models."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "*"
botocore = "*"
huggingface = "^0.0.1"
transformers = "^4.41.1"
ipykernel = "^6.29.4"
pandas = "^2.2.2"
litellm = "^1.34.6"
datasets = "^2.19.1"
torch = "^2.2.1"
torchaudio = "^2.2.1"
torchvision = "^0.17.1"
sentence-transformers = "^2.5.1"
s3fs = {extras = ["boto3"], version = ">=2023.12.0"}
faiss-cpu = "^1.8.0"
numpy = "^1.26.4"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


